[HU-7/8/14/15] Alfredo Lopez: Adaptacion al modelo unificado del equipo - claves long, DbSet singulares, estados de mascota en texto y conexion local

+++ HuellitasSV.API/Controllers/NotificacionesController.cs
@@ -31,7 +31,7 @@ public class NotificacionesController : ControllerBase
     /// <returns>Lista de notificaciones; vacía si no hay coincidencias.</returns>
     [HttpGet]
     public async Task<ActionResult<IEnumerable<Notificacion>>> GetNotificaciones(
-        [FromQuery] int? refugioId,
+        [FromQuery] long? refugioId,
         [FromQuery] int? usuarioId)
     {
         IQueryable<Notificacion> consulta = _context.Notificaciones.AsQueryable();
